refactor(web): extract error display components and enhance Dashboard visibility
- Extract PasswordIssuesBanner component for reusable password error alerts
- Extract InstanceErrorDisplay component for consistent error handling
- Refactor InstanceCard to use new InstanceErrorDisplay component
- Update Dashboard to show error states with navigation to instances page
- Add PasswordIssuesBanner to both Dashboard and Instances pages
- Improve error visibility across the application

This refactoring eliminates code duplication while adding better error
visibility to the Dashboard page, making it easier for users to identify
and resolve instance connection issues.
